# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Delivery API",
        "description": "API for Food Delivery Application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/menus": {
            "get": {
                "tags": [
                    "Menus"
                ],
                "summary": "Get all menu items",
                "operationId": "1dc6b7b84ebea40905c68ce4f844da25",
                "responses": {
                    "200": {
                        "description": "List of menu items"
                    }
                }
            },
            "post": {
                "tags": [
                    "Menus"
                ],
                "summary": "Create a new menu item",
                "operationId": "37652f848e93b047d0f6f39cb4df6269",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "restaurant_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    },
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu item created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/menus/{id}": {
            "get": {
                "tags": [
                    "Menus"
                ],
                "summary": "Get a specific menu item",
                "operationId": "358e9ead3d3f6d8fd5b51d1f0882dc83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu item details"
                    },
                    "404": {
                        "description": "Menu item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Menus"
                ],
                "summary": "Update a menu item",
                "operationId": "67ceffba16f3b7be4391a4d6681fa431",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "restaurant_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    },
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu item updated successfully"
                    },
                    "404": {
                        "description": "Menu item not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menus"
                ],
                "summary": "Delete a menu item",
                "operationId": "abff8a90938116c08c448319c3dc7323",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu item deleted successfully"
                    },
                    "404": {
                        "description": "Menu item not found"
                    }
                }
            }
        },
...
    },
...
}